pyproject.toml
[project]
name = "orion"
dynamic = ["version"]
requires-python = ">=3.11,<3.15"
dependencies = [
    "click==8.1.7",
    "setuptools_scm>=6.2",
    "apache-otava @ git+https://github.com/apache/otava.git@3b3d59a787ea03749e972c3509241114b861c99e",
    "elastic-transport==8.11.0",
    "opensearch-dsl==2.1.0",
    "opensearch-py==3.0.0",
    "Jinja2==3.1.3",
    "PyYAML==6.0.1",
    "pyshorteners==1.0.1",
    "numpy==2.2.0",
    "scikit-learn==1.5.0",
    "pandas==2.3.3",
    "tabulate==0.9.0",
    "plotly>=5.18.0",
]

[project.scripts]
orion = "main:main"

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_scheme = "no-guess-dev"
local_scheme = "dirty-tag"

[tool.pylint.format]
max-line-length = 110

